Move Entity information to different package
